fixed swerve optimize decisions
re-implemented swerve with PIDController instead of CANPIDController to take advantage of PIDController's enableContinuousInput function to do optimization.  last time we tried to implement this, we were not normalizing the angle being passed to the PIDController.  also, removed the trapezoidal motion profile for now since it wasn't easy to tune.  without that motion profile, I also experimented with PID constants for drive and angle.